Update README.md
Added a link to a zipped up tomcat deployment directory from which you can run the webapp.

If CATALINA_HOME is pointed to an unzipped copy of this directory, a "mvn clean install" from the root source directory will deploy a copy of the web app's .war file.  Also, recall that a value for NUXEO_HOME needs to be set as either an environment variable or as a Java system property and point to the "nxserver" directory inside the tomcat directory.